enable hibernate validation, fix data types

+++ db-api/src/main/kotlin/com/vsharkovski/dbpaperapi/api/PublicVariable.kt
@@ -1,7 +1,7 @@
 package com.vsharkovski.dbpaperapi.api
 
 data class PublicVariable(
-    val id: Int,
+    val id: Short,
     val name: String,
-    val type: Int? = null
+    val type: Short? = null
 )
